requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
filelock==3.12.0
google-auth==2.19.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
hydra-core==1.3.2
idna==3.4
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.0
torch==2.0.1
torch-tb-profiler==0.4.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.4
